requirements.txt
streamlit
openai
pgeocode
python-dotenv
pandas
